# ios-plan-checkin/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "计划打卡 API", "version": "0.1.0" },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "getMe",
        "tags": ["me"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "requestId", "serverTime"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/User" },
                    "requestId": { "type": "string" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateMe",
        "tags": ["me"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "X-Client-Request-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMeRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "requestId", "serverTime"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/User" },
                    "requestId": { "type": "string" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/checkins/{id}/encouragements": {
      "post": {
        "operationId": "createEncouragement",
        "tags": ["checkins"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "X-Client-Request-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEncouragementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "requestId", "serverTime"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/Encouragement" },
                    "requestId": { "type": "string" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listEncouragements",
        "tags": ["checkins"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "requestId", "serverTime"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Encouragement" }
                    },
                    "requestId": { "type": "string" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/devices/push-token": {
      "post": {
        "operationId": "registerPushToken",
        "tags": ["devices"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "X-Client-Request-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "requestId", "serverTime"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PushTokenRegistration"
                    },
                    "requestId": { "type": "string" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/notification-preferences": {
      "get": {
        "operationId": "getNotificationPreferences",
        "tags": ["me"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "requestId", "serverTime"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationPreferences"
                    },
                    "requestId": { "type": "string" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateNotificationPreferences",
        "tags": ["me"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "X-Client-Request-Id",
            "in": "header",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferencesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "requestId", "serverTime"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/NotificationPreferences"
                    },
                    "requestId": { "type": "string" },
                    "serverTime": { "type": "string", "format": "date-time" }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": ["code", "message", "requestId"],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "RATE_LIMITED",
              "OTP_EXPIRED",
              "OTP_INVALID",
              "OTP_RATE_LIMITED",
              "SMS_UNAVAILABLE",
              "USERNAME_TAKEN",
              "PHONE_TAKEN",
              "PLAN_DATE_INVALID",
              "RULE_CHANGED",
              "CHECKIN_CONFLICT",
              "PLAN_NOT_ACTIVE",
              "IDEMPOTENCY_KEY_REUSED",
              "SHARE_REVOKED",
              "FRIEND_BLOCKED",
              "MEDIA_NOT_READY",
              "CURSOR_EXPIRED",
              "ACCOUNT_DELETION_PENDING"
            ]
          },
          "message": { "type": "string" },
          "requestId": { "type": "string" },
          "details": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": false
      },
...
      "CreateEncouragementRequest": {
        "type": "object",
        "required": ["kind", "body"],
        "properties": {
          "kind": { "enum": ["emoji", "message"] },
          "body": { "type": "string", "minLength": 1, "maxLength": 500 }
        },
        "additionalProperties": false
      },
      "Encouragement": {
        "type": "object",
        "required": ["id", "checkinId", "sender", "kind", "body", "createdAt"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "checkinId": { "type": "string", "format": "uuid" },
          "sender": { "$ref": "#/components/schemas/SocialUser" },
          "kind": { "enum": ["emoji", "message"] },
          "body": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": false
      },
      "RegisterPushTokenRequest": {
        "type": "object",
        "required": ["deviceId", "platform", "token", "enabled"],
        "properties": {
          "deviceId": { "type": "string", "format": "uuid" },
          "platform": { "const": "ios" },
          "token": {
            "oneOf": [
              { "type": "string", "pattern": "^[0-9a-fA-F]{64,200}$" },
              { "type": "null" }
            ]
          },
          "enabled": { "type": "boolean" }
        },
        "additionalProperties": false
      },
      "PushTokenRegistration": {
        "type": "object",
        "required": ["deviceId", "registered"],
        "properties": {
          "deviceId": { "type": "string", "format": "uuid" },
          "registered": { "type": "boolean" }
        },
        "additionalProperties": false
      },
      "NotificationPreferences": {
        "type": "object",
        "required": [
          "friendRequests",
          "sharedUpdates",
          "encouragements",
          "revision",
          "updatedAt"
        ],
        "properties": {
          "friendRequests": { "type": "boolean" },
          "sharedUpdates": { "type": "boolean" },
          "encouragements": { "type": "boolean" },
          "revision": { "type": "integer", "minimum": 1 },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "additionalProperties": false
      },
      "UpdateNotificationPreferencesRequest": {
        "type": "object",
        "required": ["baseRevision"],
        "properties": {
          "baseRevision": { "type": "integer", "minimum": 1 },
          "friendRequests": { "type": "boolean" },
          "sharedUpdates": { "type": "boolean" },
          "encouragements": { "type": "boolean" }
        },
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "nickname",
          "avatarMediaId",
          "accountStatus",
          "revision"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "oneOf": [{ "type": "string" }, { "type": "null" }] },
          "nickname": { "oneOf": [{ "type": "string" }, { "type": "null" }] },
          "avatarMediaId": {
            "oneOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ]
          },
          "accountStatus": { "enum": ["active", "deletion_pending"] },
          "revision": { "type": "integer", "minimum": 1 }
        },
        "additionalProperties": false
      },
      "SocialUser": {
        "type": "object",
        "required": ["id", "username", "nickname", "avatarMediaId"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "nickname": { "oneOf": [{ "type": "string" }, { "type": "null" }] },
          "avatarMediaId": {
            "oneOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ]
          }
        },
        "additionalProperties": false
      },
...
      "UpdateMeRequest": {
        "type": "object",
        "required": ["baseRevision"],
        "properties": {
          "baseRevision": { "type": "integer", "minimum": 1 },
          "username": { "type": "string" },
          "nickname": { "type": "string" },
          "avatarMediaId": {
            "oneOf": [
              { "type": "string", "format": "uuid" },
              { "type": "null" }
            ]
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}